+++ ext/escape_utils/escape_utils.c
@@ -95,21 +95,6 @@ static VALUE rb_eu_unescape_js(VALUE self, VALUE str)
 	return rb_eu__generic(str, &houdini_unescape_js);
 }
 
-
-/**
- * URL methods
- */
-static VALUE rb_eu_escape_url(VALUE self, VALUE str)
-{
-	return rb_eu__generic(str, &houdini_escape_url);
-}
-
-static VALUE rb_eu_unescape_url(VALUE self, VALUE str)
-{
-	return rb_eu__generic(str, &houdini_unescape_url);
-}
-
-
 /**
  * URI methods
  */
@@ -149,8 +134,6 @@ void Init_escape_utils()
 	rb_define_method(rb_mEscapeUtils, "escape_xml", rb_eu_escape_xml, 1);
 	rb_define_method(rb_mEscapeUtils, "escape_javascript", rb_eu_escape_js, 1);
 	rb_define_method(rb_mEscapeUtils, "unescape_javascript", rb_eu_unescape_js, 1);
-	rb_define_method(rb_mEscapeUtils, "escape_url", rb_eu_escape_url, 1);
-	rb_define_method(rb_mEscapeUtils, "unescape_url", rb_eu_unescape_url, 1);
 	rb_define_method(rb_mEscapeUtils, "escape_uri", rb_eu_escape_uri, 1);
 	rb_define_method(rb_mEscapeUtils, "unescape_uri", rb_eu_unescape_uri, 1);
 	rb_define_method(rb_mEscapeUtils, "escape_uri_component", rb_eu_escape_uri_component, 1);