+++ ext/escape_utils/houdini_uri_u.c
@@ -7,13 +7,13 @@
 #define hex2c(c) ((c | 32) % 39 - 9)
 
 static int
-unescape(gh_buf *ob, const uint8_t *src, size_t size, bool unescape_plus)
+unescape(gh_buf *ob, const uint8_t *src, size_t size)
 {
 	size_t  i = 0, org;
 
 	while (i < size) {
 		org = i;
-		while (i < size && src[i] != '%' && src[i] != '+')
+		while (i < size && src[i] != '%')
 			i++;
 
 		if (likely(i > org)) {
@@ -31,11 +31,7 @@ unescape(gh_buf *ob, const uint8_t *src, size_t size, bool unescape_plus)
 		if (i >= size)
 			break;
 
-		if (src[i++] == '+') {
-			gh_buf_putc(ob, unescape_plus ? ' ' : '+');
-			continue;
-		}
-
+		i++;
 		if (i + 1 < size && _isxdigit(src[i]) && _isxdigit(src[i + 1])) {
 			unsigned char new_char = (hex2c(src[i]) << 4) + hex2c(src[i + 1]);
 			gh_buf_putc(ob, new_char);
@@ -51,18 +47,12 @@ unescape(gh_buf *ob, const uint8_t *src, size_t size, bool unescape_plus)
 int
 houdini_unescape_uri(gh_buf *ob, const uint8_t *src, size_t size)
 {
-	return unescape(ob, src, size, false);
+	return unescape(ob, src, size);
 }
 
 int
 houdini_unescape_uri_component(gh_buf *ob, const uint8_t *src, size_t size)
 {
-	return unescape(ob, src, size, false);
-}
-
-int
-houdini_unescape_url(gh_buf *ob, const uint8_t *src, size_t size)
-{
-	return unescape(ob, src, size, true);
+	return unescape(ob, src, size);
 }