+++ test/url/escape_test.rb
@@ -41,22 +41,6 @@ class UrlEscapeTest < Minitest::Test
     assert_equal "a%2Fslash", EscapeUtils.escape_url("a/slash")
   end
 
-  def test_input_must_be_utf8_or_ascii
-    str = "fo<o>bar"
-
-    str.force_encoding 'ISO-8859-1'
-    assert_raises Encoding::CompatibilityError do
-      EscapeUtils.escape_url(str)
-    end
-
-    str.force_encoding 'UTF-8'
-    begin
-      EscapeUtils.escape_url(str)
-    rescue Encoding::CompatibilityError => e
-      assert_nil e, "#{e.class.name} raised, expected not to"
-    end
-  end
-
   def test_return_value_is_tagged_as_utf8
     str = "fo<o>bar"
     assert_equal Encoding.find('UTF-8'), EscapeUtils.escape_url(str).encoding