+++ test/url/unescape_test.rb
@@ -46,22 +46,6 @@ class UrlUnescapeTest < Minitest::Test
     end
   end
 
-  def test_input_must_be_valid_utf8_or_ascii
-    escaped = EscapeUtils.escape_url("fo<o>bar")
-
-    escaped.force_encoding 'ISO-8859-1'
-    assert_raises Encoding::CompatibilityError do
-      EscapeUtils.unescape_url(escaped)
-    end
-
-    escaped.force_encoding 'UTF-8'
-    begin
-      EscapeUtils.unescape_url(escaped)
-    rescue Encoding::CompatibilityError => e
-      assert_nil e, "#{e.class.name} raised, expected not to"
-    end
-  end
-
   def test_return_value_is_tagged_as_utf8
     escaped = EscapeUtils.escape_url("fo<o>bar")
     assert_equal Encoding.find('UTF-8'), EscapeUtils.unescape_url(escaped).encoding